+++ resources/database/migrations/2016_11_12_002958_shipping_queue.php
@@ -14,8 +14,8 @@ class ShippingQueue extends Migration
     {
         Schema::create('shipping_queue', function(Blueprint $table){
             $table->increments('id');
-            $table->integer('claim_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('claim_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->timestamps();
 
             $table->foreign('claim_id')
